// frontend/src/app/api/model/settings.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DefaultView } from './defaultView';


export interface Settings { 
    displayName: string;
    inAppNotifications: boolean;
    browserNotifications: boolean;
    defaultView: DefaultView;
}
export namespace Settings {
}